% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterdown.R
\docType{package}
\name{posterdownLaTeX}
\alias{posterdownLaTeX}
\alias{posterdownLaTeX-package}
\title{posterdownLaTeX package}
\description{
This package creates posters for conferences
}
\section{posterdown_latex}{

Creates an R Markdown template for a PDF poster document via LaTeX
}

\author{
Ryan McShane
}
